Add license, readme, refactor base classes into their own files
