+++ src/main/java/industries/laser/mannequin/ValidationEvent.kt
@@ -1,0 +1,6 @@
+package industries.laser.mannequin
+
+enum class ValidationEvent {
+    OnKey,
+    OnFocusLost
+}
